requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
aiohttp=3.8.1=py38h497a2fe_0
aiosignal=1.2.0=pyhd8ed1ab_0
alembic=1.5.8=pyhd8ed1ab_0
alsa-lib=1.2.3=h516909a_0
ansi2html=1.7.0=py38h578d9bd_0
anyio=2.2.0=py38h578d9bd_0
argon2-cffi=20.1.0=py38h497a2fe_2
async-timeout=4.0.2=pyhd8ed1ab_0
async_generator=1.10=py_0
atk-1.0=2.36.0=h3371d22_4
attrs=20.3.0=pyhd3deb0d_0
babel=2.9.0=pyhd3deb0d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.3=pyhd8ed1ab_0
bleach=3.3.0=pyh44b312d_0
blinker=1.4=py_1
brotli-python=1.0.9=py38h709712a_6
brotlipy=0.7.0=py38h497a2fe_1001
bzip2=1.0.8=h7f98852_4
c-ares=1.17.1=h7f98852_1
ca-certificates=2021.10.8=ha878542_0
cairo=1.16.0=h6cf1ce9_1008
certifi=2021.10.8=py38h578d9bd_1
certipy=0.1.3=py_0
cffi=1.15.0=py38h3931269_0
chardet=4.0.0=py38h578d9bd_1
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=8.0.4=py38h578d9bd_0
cmdstanpy=1.0.2=pypi_0
conda=4.10.0=py38h578d9bd_1
conda-package-handling=1.7.2=py38h8df0ef7_0
configurable-http-proxy=4.3.1=node15_he6ea98c_0
convertdate=2.4.0=pypi_0
cryptography=3.4.7=py38ha5dfef3_0
cython=0.29.30=pypi_0
dash=2.3.0=pyhd8ed1ab_0
dash-labs=1.1.0=pypi_0
dbus=1.13.6=h48d8840_2
decorator=5.0.6=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.3=pyhd8ed1ab_1003
ephem=4.1.3=pypi_0
expat=2.4.1=h9c3ff4c_0
flask=1.1.2=pyh9f0ad1d_0
flask-compress=1.11=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.13.1=hba837de_1005
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freetype=2.10.4=h0708190_1
fribidi=1.0.10=h36c2ea0_0
frozenlist=1.3.0=py38h497a2fe_0
fuzzywuzzy=0.18.0=pyhd8ed1ab_0
gdk-pixbuf=2.42.8=hff1cb4f_0
gettext=0.19.8.1=h73d1719_1008
giflib=5.2.1=h36c2ea0_2
glib=2.70.2=h780b84a_4
glib-tools=2.70.2=h780b84a_4
graphite2=1.3.13=h58526e2_1001
graphviz=2.50.0=h85b4f2f_1
greenlet=1.0.0=py38h709712a_0
gst-plugins-base=1.18.4=hf529b03_2
gstreamer=1.18.4=h76c114f_2
gtk2=2.24.33=h539f30e_1
gts=0.7.6=h64030ff_2
harfbuzz=2.9.1=h83ec7ef_1
hijri-converter=2.2.4=pypi_0
holidays=0.14.2=pypi_0
icu=68.1=h58526e2_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=3.10.0=py38h578d9bd_0
ipykernel=5.5.3=py38hd0cf306_0
ipython=7.22.0=py38hd0cf306_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.7.0=pyhd8ed1ab_0
itsdangerous=2.1.1=pyhd8ed1ab_0
jbig=2.1=h7f98852_2003
jedi=0.18.0=py38h578d9bd_2
jinja2=2.11.3=pyh44b312d_0
jpeg=9e=h7f98852_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter=1.0.0=py38h578d9bd_7
jupyter-dash=0.4.1=pyhd8ed1ab_2
jupyter-packaging=0.7.12=pyhd8ed1ab_0
jupyter-server-proxy=3.2.1=pyhd8ed1ab_0
jupyter_client=6.1.12=pyhd8ed1ab_0
jupyter_console=6.4.2=pyhd8ed1ab_0
jupyter_core=4.7.1=py38h578d9bd_0
jupyter_server=1.5.1=py38h578d9bd_0
jupyter_telemetry=0.1.0=pyhd8ed1ab_1
jupyterhub=1.3.0=py38h578d9bd_1
jupyterhub-base=1.3.0=py38h578d9bd_1
jupyterlab=3.0.12=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=2.4.0=pyhd8ed1ab_0
jupyterlab_widgets=1.1.0=pyhd8ed1ab_0
korean-lunar-calendar=0.2.1=pypi_0
krb5=1.17.2=h926e7f8_0
ld_impl_linux-64=2.35.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libarchive=3.5.2=hccf745f_1
libblas=3.9.0=13_linux64_openblas
libcblas=3.9.0=13_linux64_openblas
libclang=11.1.0=default_ha53f305_1
libcurl=7.76.0=hc4aaa36_0
libdeflate=1.10=h7f98852_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=hcdb4288_3
libffi=3.4.2=h7f98852_5
libgcc-ng=11.2.0=h1d223b6_14
libgd=2.3.3=h6ad9fb6_0
libgfortran-ng=11.2.0=h69a702a_14
libgfortran5=11.2.0=h5c6108e_14
libglib=2.70.2=h174f98d_4
libgomp=11.2.0=h1d223b6_14
libiconv=1.16=h516909a_0
liblapack=3.9.0=13_linux64_openblas
libllvm11=11.1.0=hf817b99_2
libnghttp2=1.43.0=h812cca2_0
libnsl=2.0.0=h7f98852_0
libogg=1.3.4=h7f98852_1
libopenblas=0.3.18=pthreads_h8fe5266_0
libopus=1.3.1=h7f98852_1
libpng=1.6.37=h21135ba_2
libpq=13.2=hfd2b0eb_2
librsvg=2.52.5=hc3c00ef_1
libsodium=1.0.18=h36c2ea0_1
libsolv=0.7.18=h780b84a_0
libssh2=1.9.0=ha56f1ee_6
libstdcxx-ng=11.2.0=he4da1e4_14
libtiff=4.3.0=h542a066_3
libtool=2.4.6=h9c3ff4c_1008
libuuid=2.32.1=h7f98852_1000
libuv=1.41.0=h7f98852_0
libvorbis=1.3.7=h9c3ff4c_0
libwebp=1.2.2=h3452ae3_0
libwebp-base=1.2.2=h7f98852_1
libxcb=1.13=h7f98852_1003
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h72842e0_0
libzlib=1.2.11=h36c2ea0_1013
lunarcalendar=0.0.9=pypi_0
lz4-c=1.9.3=h9c3ff4c_0
lzo=2.10=h516909a_1000
mako=1.1.4=pyh44b312d_0
mamba=0.9.2=py38h2aa5da1_0
markupsafe=1.1.1=py38h497a2fe_3
mistune=0.8.4=py38h497a2fe_1003
multidict=6.0.2=py38h497a2fe_0
mysql-common=8.0.28=ha770c72_0
mysql-libs=8.0.28=hfa10184_0
nbclassic=0.2.6=pyhd8ed1ab_0
nbclient=0.5.3=pyhd8ed1ab_0
nbconvert=6.0.7=py38h578d9bd_3
nbformat=5.1.3=pyhd8ed1ab_0
nbgrader=0.6.1=py38h578d9bd_1
ncurses=6.3=h9c3ff4c_0
nest-asyncio=1.5.1=pyhd8ed1ab_0
nodejs=15.14.0=h92b4a50_0
notebook=6.3.0=py38h578d9bd_0
nspr=4.30=h9c3ff4c_0
nss=3.64=hb5efdd6_0
numpy=1.22.3=py38h05e7239_0
oauthlib=3.0.1=py_0
openssl=1.1.1l=h7f98852_0
packaging=20.9=pyh44b312d_0
pamela=1.0.0=py_0
pandas=1.4.1=py38h43a58ef_0
pandoc=2.12=h7f98852_0
pandocfilters=1.4.2=py_1
pango=1.48.10=hb8ff022_1
parso=0.8.2=pyhd8ed1ab_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pip=21.0.1=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
plotly=5.6.0=pyhd8ed1ab_0
prometheus_client=0.10.0=pyhd8ed1ab_0
prompt-toolkit=3.0.18=pyha770c72_0
prompt_toolkit=3.0.18=hd8ed1ab_0
prophet=1.1=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pycosat=0.6.3=py38h497a2fe_1006
pycparser=2.20=pyh9f0ad1d_2
pycurl=7.43.0.6=py38h996a351_1
pygments=2.8.1=pyhd8ed1ab_0
pyjwt=2.0.1=pyhd8ed1ab_1
pymeeus=0.5.11=pypi_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py38h578d9bd_7
pyqt-impl=5.12.3=py38h7400c14_7
pyqt5-sip=4.19.18=py38h709712a_7
pyqtchart=5.12=py38h7400c14_7
pyqtwebengine=5.12.1=py38h7400c14_7
pyrsistent=0.17.3=py38h497a2fe_2
pysocks=1.7.1=py38h578d9bd_3
python=3.8.12=ha38a3c6_3_cpython
python-dateutil=2.8.1=py_0
python-editor=1.0.4=py_0
python-json-logger=2.0.1=pyh9f0ad1d_0
python-levenshtein=0.12.2=py38h497a2fe_0
python_abi=3.8=1_cp38
pytz=2021.1=pyhd8ed1ab_0
pyzmq=22.0.3=py38h2035c66_1
qt=5.12.9=hda022c4_4
qtconsole=5.2.2=pyhd8ed1ab_1
qtconsole-base=5.2.2=pyhd8ed1ab_1
qtpy=2.0.1=pyhd8ed1ab_0
readline=8.1=h46c0cb4_0
reproc=14.2.1=h36c2ea0_0
reproc-cpp=14.2.1=h58526e2_0
requests=2.25.1=pyhd3deb0d_0
retrying=1.3.3=py_2
ruamel.yaml=0.16.12=py38h497a2fe_2
ruamel.yaml.clib=0.2.2=py38h497a2fe_2
ruamel_yaml=0.15.80=py38h497a2fe_1004
send2trash=1.5.0=py_0
setuptools=49.6.0=py38h578d9bd_3
setuptools-git=1.2=pypi_0
simpervisor=0.4=pyhd8ed1ab_0
six=1.15.0=pyh9f0ad1d_0
sniffio=1.2.0=py38h578d9bd_1
sqlalchemy=1.4.6=py38h497a2fe_0
sqlite=3.37.1=h4ff8645_0
tenacity=8.0.1=pyhd8ed1ab_0
terminado=0.9.4=py38h578d9bd_0
testpath=0.4.4=py_0
tini=0.18.0=h14c3975_1001
tk=8.6.12=h27826a3_0
tornado=6.1=py38h497a2fe_1
tqdm=4.60.0=pyhd8ed1ab_0
traitlets=5.0.5=py_0
typing-extensions=4.1.1=hd8ed1ab_0
typing_extensions=4.1.1=pyha770c72_0
ujson=5.3.0=pypi_0
urllib3=1.26.4=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=2.0.3=pyhd8ed1ab_1
wheel=0.37.1=pypi_0
widgetsnbextension=3.6.0=py38h578d9bd_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
yarl=1.7.2=py38h497a2fe_1
zeromq=4.3.4=h9c3ff4c_0
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h36c2ea0_1013
zstd=1.5.2=ha95c52a_0
